REFACTOR - extracted PageEntryWidget, CommentEntryWidget and ParameterEntryWidget into their own files in the widget/
